+++ projects/developer/src/app/app-routing.module.ts
@@ -15,8 +15,9 @@ import { MetricsComponent } from './data/metrics/component';
 import { TimelineComponent } from './data/timeline/component';
 import { RunningJobsComponent } from './processing/running-jobs/component';
 import { QueuedJobsComponent } from './processing/queued-jobs/component';
-import { BatchesComponent } from './processing/batches/component';
+import { BatchesComponent } from './processing/batches/batches-component';
 import { BatchDetailsComponent } from './processing/batches/details.component';
+import { BatchWorkflowComponent } from './processing/batches/batch-workflow/batch-workflow.component';
 import { IngestComponent } from './data/ingest/component';
 import { FeedComponent } from './data/feed/component';
 import { NodesComponent } from './system/nodes/component';
@@ -74,17 +75,16 @@ const routes: Routes = [
         data: {title: 'Batches | Scale'}
     },
     {
+        path: 'processing/batches/create',
+        component: BatchWorkflowComponent,
+        data: {title: 'Create Batch | Scale'}
+    },
+    {
         path: 'processing/batches/:id',
         component: BatchDetailsComponent,
         canDeactivate: [PendingChangesGuard],
         data: {title: 'Batch Details | Scale'}
     },
-    // {
-    //     path: 'processing/batches/create',
-    //     component: BatchDetailsComponent,
-    //   //  canDeactivate: [PendingChangesGuard],
-    //     data: {title: 'Create Batch | Scale'}
-    // },
     {
         path: 'configuration/job-types',
         component: JobTypesComponent,