+++ projects/developer/src/app/app.module.ts
@@ -5,6 +5,7 @@ import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 import { HttpClientModule } from '@angular/common/http';
 import { HttpClientXsrfModule } from '@angular/common/http';
+import { PlotlyViaWindowModule } from 'angular-plotly.js';
 
 // vendor
 import { CodemirrorModule } from '@ctrl/ngx-codemirror';
@@ -12,13 +13,14 @@ import { NgxGraphModule } from '@swimlane/ngx-graph';
 import { SeedImagesModule } from '@ngageoint/seed-images';
 import { UtcDatepickerModule } from 'angular-utc-datepicker';
 
+
 // Prime NG
 import {
     AutoCompleteModule, ButtonModule, CalendarModule, ChartModule, CheckboxModule, ChipsModule, ColorPickerModule, ConfirmationService,
-    DataListModule, DialogModule, DropdownModule, InputSwitchModule, InputTextModule, InputTextareaModule, ListboxModule, MenubarModule,
-    MessageModule, MessagesModule, MultiSelectModule, OverlayPanelModule, PaginatorModule, PanelModule, ScrollPanelModule,
+    DataListModule, DialogModule, DropdownModule, FieldsetModule, InputSwitchModule, InputTextModule, InputTextareaModule, ListboxModule,
+    MenubarModule, MessageModule, MessagesModule, MultiSelectModule, OverlayPanelModule, PaginatorModule, PanelModule, ScrollPanelModule,
     SelectButtonModule, SidebarModule, SliderModule, SpinnerModule, StepsModule, TabViewModule, ToggleButtonModule, TooltipModule,
-    TreeTableModule, SlideMenuModule
+    TreeTableModule, SlideMenuModule, Fieldset
 } from 'primeng/primeng';
 import { AccordionModule } from 'primeng/accordion';
 import { CardModule } from 'primeng/card';
@@ -36,7 +38,7 @@ import { AppComponent } from './app.component';
 import { AppConfigService } from './common/services/app-config.service';
 import { AppRoutingModule } from './app-routing.module';
 import { BatchDetailsComponent } from './processing/batches/details.component';
-import { BatchesComponent } from './processing/batches/component';
+import { BatchesComponent } from './processing/batches/batches-component';
 import { CandyBarComponent } from './common/components/candy-bar/component';
 import { DashboardComponent } from './dashboard/component';
 import { DataFeedComponent } from './dashboard/data-feed/component';
@@ -81,6 +83,10 @@ import { TemporalFilterComponent } from './common/components/temporal-filter/com
 import { ThemeModule, lightTheme, darkTheme } from './theme';
 import { TimelineComponent } from './data/timeline/component';
 import { WorkspacesComponent } from './system/workspaces/component';
+import { BatchWorkflowComponent } from './processing/batches/batch-workflow/batch-workflow.component';
+import { CreateDatasetComponent } from './processing/batches/batch-workflow/create-dataset/create-dataset.component';
+import { CreateBatchComponent } from './processing/batches/batch-workflow/create-batch/create-batch.component';
+import { RunBatchComponent } from './processing/batches/batch-workflow/run-batch/run-batch.component';
 
 const appInitializer = (appConfig: AppConfigService) => {
     return () => {
@@ -138,7 +144,11 @@ const appInitializer = (appConfig: AppConfigService) => {
         SystemStatusComponent,
         TimelineComponent,
         WorkspacesComponent,
-        TemporalFilterComponent
+        TemporalFilterComponent,
+        BatchWorkflowComponent,
+        CreateDatasetComponent,
+        CreateBatchComponent,
+        RunBatchComponent
     ],
     imports: [
         AccordionModule,
@@ -159,6 +169,7 @@ const appInitializer = (appConfig: AppConfigService) => {
         DataViewModule,
         DialogModule,
         DropdownModule,
+        FieldsetModule,
         HttpClientModule,
         HttpClientXsrfModule.withOptions({
             cookieName: 'csrftoken',
@@ -178,6 +189,7 @@ const appInitializer = (appConfig: AppConfigService) => {
         OverlayPanelModule,
         PaginatorModule,
         PanelModule,
+        PlotlyViaWindowModule,
         ProgressBarModule,
         ProgressSpinnerModule,
         ReactiveFormsModule,